+++ client/index.android.js
@@ -1,6 +1,6 @@
 'use strict';
 
 import React, { AppRegistry } from 'react-native';
-import App from './src/containers/app';
+import App from './src/views/app';
 
 AppRegistry.registerComponent('trailangel', () => App);
