+++ client/src/views/favorites.js
